donor/AgentKB_MCP/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
pgvector==0.2.4

# Redis
redis[hiredis]==5.0.1

# Google AI Services
google-generativeai==0.3.0
google-cloud-storage==2.10.0

# Fallback AI Services
anthropic==0.7.0
qdrant-client==1.7.0

# HTTP Client
httpx==0.25.2

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0

# Circuit Breaker
pybreaker==1.0.1

# Multipart Form Data
python-multipart==0.0.6

# Scientific Computing (for semantic dedup)
numpy==1.24.3
scikit-learn==1.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# YAML for golden dataset
pyyaml==6.0.1

# CLI
click==8.1.7
rich==13.7.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

